Add multiarch for fbc 4.17 pipelineRun in pull
This commit modifies the 4.17 pull-request pipelineRun to trigger multi
arch builds. It also fixes a missing change in the push pipelineRun

It changes the git-cone and buildah to the oci-ta variants while also
still using a workspace due to the implementation of the fbc checks. It
uses [1] as a sample.

[1] https://github.com/konflux-ci/olm-operator-konflux-sample/pull/65/files#
Signed-off-by: Daniel Mellado <dmellado@fedoraproject.org>