Build fbc catalog only on change in pipelines.
This commit modifies the trigger for the catalog build so it applies
only when there's a change in the actual catalog in order to avoid
having it rebuilt on every commit.

Note that this would mean stopping seeing this in every regular commit
CI, but that wasn't a real update but a rebuilt.  Currently there's no
automation in place to have the catalogs updated per commit in the CI.

Signed-off-by: Daniel Mellado <dmellado@fedoraproject.org>