Accomdate usage of tags instead of branches
Add /perses-operator to .gitignore
Add `ignore=all` flag to .gitmodules > submodule "perses-operator" so that it doesn't update the submodule when runnning `git submodule updates`.

We'll need to remove the gitignore and manually push for each tag change to point at a new commit in the submodule.

(current workaround the fact that there are no release branches in perses/perses-operator)