Use proper registry for FBC
This PR modifies the pullspec in the catalog for the FBCs while also
introducing a modification on the make-generate Makefile target.

This ensures pullspec validity when running the opm command and
post-processes the catalog to update the pullspec later.

This will also make EC/Conforma validation fails, it's a known issue.

Signed-off-by: Daniel Mellado <dmellado@fedoraproject.org>